Update README.md
Hey @gandm, I think this is a very important package in the react community. If you would like to pass this project on instead of not supporting it anymore, I would like to offer to maintain it. The only way I could get in contact with you is by opening a PR. Please contact me here or via email (michael@mgmcdermott.com) if you would consider letting me take over. Thanks